chore: Separate classes and extensions into separate files
